+++ CHANGELOG.md
@@ -21,6 +21,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Support for displaying custom images in case view
 - Individual-specific HPO terms
 - Optional alamut_key in institute settings for Alamut Plus software
+- Case report API endpoint
 ### Fixed
 - Updated IGV to v2.8.5 to solve missing gene labels on some zoom levels
 - Demo cancer case config file to load somatic SNVs and SVs only.